fix error on argument type

+++ scripts/train.hierarchical_with_cycle.py
@@ -528,7 +528,7 @@ class Encoder(nn.Module):
             BNSNConv2dDBlock(16, 64, ks, 2, 8),
         ]
         # (input_size / 80) to make channel scalable with feat_dim/input_size .
-        blocks1d = [BNSNConv1dDBlock(64 * 10 * (input_size / 80), mfd, 3, 1)]
+        blocks1d = [BNSNConv1dDBlock(64 * 10 * int(input_size / 80), mfd, 3, 1)]
         for sf in z_scale_factors:
             blocks1d.append(StridedBNSNConv1dDBlock(mfd, mfd, ks, sf))
